Gives a better tradeoff between cpu and allocations than #65 so I am closing that PR in favour of this one.

Adds CborReader and CborWriter that contain a small buffer for optimizing CBOR header reads and writes. In contrast to #65 no pool is used for header buffers, instead the reader or writer reuses a single buffer over multiple operations. A package level pool is used for 8k buffers used when reading strings. The benchmarks show an increase in cpu but allocations are substantially reduced.

Benchstat comparison with master:
```
name               old time/op    new time/op    delta
Marshaling-8          564ns ± 1%     840ns ± 0%  +49.06%  (p=0.000 n=9+9)
Unmarshaling-8       2.75µs ± 5%    3.07µs ± 9%  +11.71%  (p=0.000 n=10+10)
LinkScan-8            744ns ± 0%     932ns ± 1%  +25.26%  (p=0.000 n=10+10)
Deferred-8           1.68µs ± 1%    1.89µs ± 2%  +12.57%  (p=0.000 n=10+10)
MapMarshaling-8       317ns ± 0%     441ns ± 0%  +39.21%  (p=0.000 n=9+9)
MapUnmarshaling-8    2.71µs ±10%    3.07µs ±12%  +13.50%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
Marshaling-8           160B ± 0%       64B ± 0%  -60.00%  (p=0.000 n=10+10)
Unmarshaling-8       3.44kB ± 0%    2.03kB ± 0%  -41.08%  (p=0.000 n=10+10)
LinkScan-8             112B ± 0%      112B ± 0%     ~     (all equal)
Deferred-8            88.0B ± 0%     88.0B ± 0%     ~     (all equal)
MapMarshaling-8       48.0B ± 0%     64.0B ± 0%  +33.33%  (p=0.000 n=10+10)
MapUnmarshaling-8    2.53kB ± 0%    1.60kB ± 0%  -36.63%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
Marshaling-8           10.0 ± 0%       2.0 ± 0%  -80.00%  (p=0.000 n=10+10)
Unmarshaling-8         43.0 ± 0%      23.0 ± 0%  -46.51%  (p=0.000 n=10+10)
LinkScan-8             1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Deferred-8             3.00 ± 0%      3.00 ± 0%     ~     (all equal)
MapMarshaling-8        5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
MapUnmarshaling-8      56.0 ± 0%      29.0 ± 0%  -48.21%  (p=0.000 n=10+10)
```

There's probably some additional opportunities for refactoring around the reader/writers but I'm leaviing this change focussed on the initial performance improvements.