@iand 2022-03-23
Updated comparison with master shows a much better time delta for the Marshaling benchmark

```
name               old time/op    new time/op    delta
Marshaling-8          564ns ± 1%     634ns ± 1%  +12.48%  (p=0.000 n=9+10)
Unmarshaling-8       2.75µs ± 5%    3.21µs ±16%  +16.98%  (p=0.000 n=10+10)
LinkScan-8            744ns ± 0%     948ns ± 3%  +27.32%  (p=0.000 n=10+10)
Deferred-8           1.68µs ± 1%    1.89µs ± 2%  +12.68%  (p=0.000 n=10+9)
MapMarshaling-8       317ns ± 0%     438ns ± 0%  +38.38%  (p=0.000 n=9+10)
MapUnmarshaling-8    2.71µs ±10%    3.18µs ±15%  +17.66%  (p=0.000 n=10+9)

name               old alloc/op   new alloc/op   delta
Marshaling-8           160B ± 0%       64B ± 0%  -60.00%  (p=0.000 n=10+10)
Unmarshaling-8       3.44kB ± 0%    2.03kB ± 0%  -41.08%  (p=0.000 n=10+9)
LinkScan-8             112B ± 0%      112B ± 0%     ~     (all equal)
Deferred-8            88.0B ± 0%     88.0B ± 0%     ~     (all equal)
MapMarshaling-8       48.0B ± 0%     64.0B ± 0%  +33.33%  (p=0.000 n=10+10)
MapUnmarshaling-8    2.53kB ± 0%    1.60kB ± 0%  -36.63%  (p=0.000 n=10+8)

name               old allocs/op  new allocs/op  delta
Marshaling-8           10.0 ± 0%       2.0 ± 0%  -80.00%  (p=0.000 n=10+10)
Unmarshaling-8         43.0 ± 0%      23.0 ± 0%  -46.51%  (p=0.000 n=10+10)
LinkScan-8             1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Deferred-8             3.00 ± 0%      3.00 ± 0%     ~     (all equal)
MapMarshaling-8        5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
MapUnmarshaling-8      56.0 ± 0%      29.0 ± 0%  -48.21%  (p=0.000 n=10+10)
```