@Stebalien 2022-03-23
Also, I'm seeing different perf numbers when comparing to master:

```
name               old time/op    new time/op    delta
Marshaling-8          497ns ± 2%     501ns ± 1%     ~     (p=0.310 n=5+5)
Unmarshaling-8       3.55µs ±15%    2.88µs ± 4%  -18.95%  (p=0.008 n=5+5)
LinkScan-8            547ns ± 2%     729ns ± 1%  +33.13%  (p=0.008 n=5+5)
Deferred-8           1.30µs ± 2%    1.49µs ± 2%  +14.80%  (p=0.008 n=5+5)
MapMarshaling-8       281ns ± 2%     348ns ± 2%  +23.88%  (p=0.008 n=5+5)
MapUnmarshaling-8    3.10µs ±13%    2.51µs ± 4%  -19.03%  (p=0.008 n=5+5)
```

What version of go are you using? (I'm using 1.18, on an intel i7 laptop).